+++ model_bakery/recipe.py
@@ -1,8 +1,9 @@
 import inspect
 import itertools
-from typing import Any, Dict, List, Union, cast
+from typing import Any, Dict, List, Type, Union, cast
 
 from django.db.models import Model
+from django.db.models.base import ModelBase
 
 from . import baker
 from .exceptions import RecipeNotFound
@@ -12,7 +13,7 @@ finder = baker.ModelFinder()
 
 
 class Recipe(object):
-    def __init__(self, _model: str, **attrs) -> None:
+    def __init__(self, _model: Union[str, Type[ModelBase]], **attrs) -> None:
         self.attr_mapping = attrs
         self._model = _model
         # _iterator_backups will hold values of the form (backup_iterator, usable_iterator).