+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `model_bakery` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/model-bakers/model_bakery/tree/master)
+## [Unreleased](https://github.com/model-bakers/model_bakery/tree/main)
 
 ### Added